Refactor tests, expose apply/generate entries
