+++ index.js
@@ -17,38 +17,5 @@ module.exports = {
   flatTop: require('./flat-top'),
   cosine: require('./cosine'),
   gaussian: require('./gaussian'),
-  tukey: require('./tukey'),
-  applyWindow: applyWindow
+  tukey: require('./tukey')
 }
-
-function applyWindow(signal, func) {
-  var i, n=signal.length, args=[0,n]
-
-  for(i=2; i<arguments.length; i++) {
-    args[i] = arguments[i]
-  }
-
-  for(i=n-1; i>=0; i--) {
-    args[0] = i
-    signal[i] *= func.apply(null,args)
-  }
-
-  return signal;
-}
-
-/*
-function generate(func, N) {
-  var i, args=[0,N]
-  var signal = new Array(N);
-
-  for(i=2; i<arguments.length; i++) {
-    args[i] = arguments[i]
-  }
-
-  for(i=N-1; i>=0; i--) {
-    args[0] = i
-    signal[i] = func.apply(null,args)
-  }
-  return signal;
-}*/
-
